+++ app.js
@@ -513,6 +513,24 @@ import commandPostObject from './commandPostObject';
 //   }
 // }
 
+// ex- TransformModifiersOnEquipment('.newSVG svg');
+// This should only be called on equipment symbols. This will scale down the decorator, and move Mod1 up and Mod2 down so they all fit in the circle
+async function TransformModifiersOnEquipment(location) {
+  if (await window.hasOwnProperty('MainMS')) {
+    const equipmentOutline = document.querySelector(location);
+    const equipmentDecorator = equipmentOutline.querySelector('g.decorator');
+    const mod1 = equipmentOutline.querySelector('g.mod1');
+    const mod2 = equipmentOutline.querySelector('g.mod2');
+    equipmentDecorator.style.transformOrigin = '100px 100px'; // transform from center of circle (cx, cy)
+    equipmentDecorator.style.transform = 'translateY(2%) scale(0.75)';
+    // mod1.style.transform = `translateY(-${equipmentOutline.viewBox.baseVal.x / equipmentOutline.viewBox.baseVal.y * 21}px)`;
+    mod1.style.transformOrigin = '100px 100px';
+    mod1.style.transform = 'translateY(-11%) scale(0.85)';
+    mod2.style.transformOrigin = '100px 140px';
+    mod2.style.transform = 'scale(0.85)';
+  }
+}
+
 // Example 1- const symbol_1 = new MilSym('.test', 'Unmanned Aerial Surveillance', 'friendlyTemplated', 'team', 'Assault', 'Rail', 'A/2-101', '27/42ID', '+', false, true, true, true, 'Main Command Post');
 // Example 2 - const symbol_2 = new MilSym('.test', 'Weapon System - Grenade Launcher - Medium', 'friendly' );
 // symbol_2.mod1 = 'Armored';
@@ -671,6 +689,10 @@ class MilSym {
         this.higherFormation = undefined;
         // Removes the Reinforced/Reduced above the Equipment symbol
         this.reinforcedReduced = undefined;
+        // setTimeout(() => {
+        TransformModifiersOnEquipment('.newSVG svg');
+        // }, 1000);
+
 
         if (this._affiliation === 'friendly') {
           const outline0 = document.createElementNS('http://www.w3.org/2000/svg', 'circle');
@@ -1103,6 +1125,7 @@ class MilSym {
     this._commandPost = value;
   }
 
+
   // PLACE SYMBOL
   placeSymbol() {
     this.location.querySelector('svg') ? this.location.querySelector('svg').remove() : null;
@@ -1256,23 +1279,24 @@ class Resizer {
 
 // ex- "new TransformModifiersOnEquipment('.newSVG > svg')"
 // This should only be called on equipment symbols. This will scale down the decorator, and move Mod1 up and Mod2 down so they all fit in the circle
-class TransformModifiersOnEquipment {
-  constructor(equipmentOutline) {
-    if (MainMS.type === 'Equipment') {
-      this.equipmentOutline = document.querySelector(equipmentOutline); // The equipment SVG you want to readjust
-      this.equipmentDecorator = this.equipmentOutline.querySelector('g.decorator');
-      this.mod1 = this.equipmentOutline.querySelector('g.mod1');
-      this.mod2 = this.equipmentOutline.querySelector('g.mod2');
-      this.equipmentDecorator.style.transformOrigin = '100px 100px'; // transform from center of circle (cx, cy)
-      this.equipmentDecorator.style.transform = 'translateY(2%) scale(0.75)';
-      // mod1.style.transform = `translateY(-${equipmentOutline.viewBox.baseVal.x / equipmentOutline.viewBox.baseVal.y * 21}px)`;
-      this.mod1.style.transformOrigin = '100px 100px';
-      this.mod1.style.transform = 'translateY(-11%) scale(0.85)';
-      this.mod2.style.transformOrigin = '100px 140px';
-      this.mod2.style.transform = 'scale(0.85)';
-    }
-  }
-}
+// class TransformModifiersOnEquipment {
+//   constructor(equipmentOutline) {
+//     if (MainMS.type === 'Equipment') {
+//       this.equipmentOutline = document.querySelector(equipmentOutline); // The equipment SVG you want to readjust
+//       this.equipmentDecorator = this.equipmentOutline.querySelector('g.decorator');
+//       this.mod1 = this.equipmentOutline.querySelector('g.mod1');
+//       this.mod2 = this.equipmentOutline.querySelector('g.mod2');
+//       this.equipmentDecorator.style.transformOrigin = '100px 100px'; // transform from center of circle (cx, cy)
+//       this.equipmentDecorator.style.transform = 'translateY(2%) scale(0.75)';
+//       // mod1.style.transform = `translateY(-${equipmentOutline.viewBox.baseVal.x / equipmentOutline.viewBox.baseVal.y * 21}px)`;
+//       this.mod1.style.transformOrigin = '100px 100px';
+//       this.mod1.style.transform = 'translateY(-11%) scale(0.85)';
+//       this.mod2.style.transformOrigin = '100px 140px';
+//       this.mod2.style.transform = 'scale(0.85)';
+//     }
+//   }
+// }
+
 
 // Toggle the bounceIn animation on the Unit Size, Mod 1 and Mod 2. Disabled for Equipment because TransformModifiersOnEquipment throws everything off
 function bounceInAnimation(location) {